UIKit specific extension of Item
