+++ EasyPeasy/Item.swift
@@ -8,7 +8,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+#if os(iOS) || os(tvOS)
 import UIKit
+#else
+import AppKit
+#endif
 
 internal var easy_attributesReference: Int = 0
 
@@ -41,50 +45,6 @@ internal extension Item {
         set {
             let policy = objc_AssociationPolicy.OBJC_ASSOCIATION_RETAIN_NONATOMIC
             objc_setAssociatedObject(self, &easy_attributesReference, newValue, policy)
-        }
-    }
-    
-    /// Owning `UIView` for the current `Item`. The concept varies
-    /// depending on the class conforming the protocol
-    internal var owningView: UIView? {
-        get {
-            // Owning view for `UIView` is the `superview`
-            if let view = self as? UIView {
-                return view.superview
-            }
-            
-            // Owning view for `UILayoutGuide` is the `owningView`
-            // iOS 9 and above only
-            if #available(iOS 9.0, *) {
-                if let layoutGuide = self as? UILayoutGuide {
-                    return layoutGuide.owningView
-                }
-            }
-            
-            return nil
-        }
-    }
-    
-}
-
-/**
-    Extension making `UIView` conform the `Item` protocol and
-    therefore inherit the extended methods and properties
- */
-extension UIView: Item { }
-
-/**
-    Extension making `UILayoutGuide` conform the `Item` protocol 
-    therefore and inherit the extended methods and properties
- */
-@available(iOS 9.0, *)
-extension UILayoutGuide: Item {
-    
-    /// Constraints in `owningView` with the current `UILayoutGuide`
-    /// as `firstItem`
-    public var constraints: [NSLayoutConstraint] {
-        get {
-            return self.owningView?.constraints.filter { $0.firstItem === self } ?? []
         }
     }
